cytron-test: Change PWM to Timer 2
This is what the speedy-breakout uses for the dedicated motor outputs.
I now realize that this design decision may have been a mistake, as the
system timer is now a 16-bit timer instead of the 32-bit Timer 2.
Potential workaround: add a bodge wire between another Timer's output
and the PA1/3 breakout connections and set PA1/3 to high impedance.

PA3 also seems to have a DC offset due to the USART RX connection with
the ST-LINK MCU. It's fine when powered off the 5V pin (such as from the
CAN bus) but when the ST-Link gets powered up when plugging the Speedy
in over USB, that pin activates.

Potential workaround: Modify the ST-Link MCY firmware to disable the
USART?