common: Remove UAVCAN related code
We have since dropped UAVCAN as the code would often break requiring a
lot of time to debug. It also introduced a level of complexity that was
unnecessary for our purposes, which also makes things harder for new
members. Currently, we are using the bare ChibiOS CAN driver and our own
custom message formats on top of the raw CAN frames. The frame format is
somewhat inspired by UAVCAN though.